tests: separated Course static and proto methods for readability
